detailed_requirements.txt
alignn==2022.1.25
amqp==5.1.1
anyio==3.6.2
asgiref==3.5.2
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
billiard==3.6.4.0
black==22.10.0
blinker==1.4
blob-utils==1.14.0
celery==5.2.3
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.6
core-composer-app==1.21.0
core-curate-app==1.21.0
core-dashboard-app==1.21.0
core-dashboard-common-app==1.21.0
core-explore-common-app==1.21.0
core-explore-example-app==1.21.0
core-explore-federated-search-app==1.21.0
core-explore-keyword-app==1.21.0
core-explore-oaipmh-app==1.21.0
core-exporters-app==1.21.0
core-federated-search-app==1.21.0
core-file-preview-app==1.19.0
core-linked-records-app==1.4.1
core-main-app==1.21.4
core-module-advanced-blob-host-app==1.21.0
core-module-blob-host-app==1.21.0
core-module-chemical-composition-app==1.21.0
core-module-chemical-composition-simple-app==1.21.0
core-module-excel-uploader-app==1.21.0
core-module-periodic-table-app==1.21.0
core-module-remote-blob-host-app==1.21.0
core-module-text-area-app==1.21.0
core-oaipmh-common-app==1.21.0
core-oaipmh-harvester-app==1.21.0
core-oaipmh-provider-app==1.21.0
core-parser-app==1.21.1
core-website-app==1.21.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
dgl==0.9.1
Django==3.2.16
django-allauth==0.46.0
django-celery-beat==2.2.1
django-crispy-forms==1.9.2
django-defender==0.9.1
django-extensions==3.2.1
django-filter==22.1
django-jsonfield==1.4.1
django-mongoengine==0.4.6
django-oauth-toolkit==1.4.1
django-ranged-response==0.2.0
django-rest-framework-mongoengine==3.4.1
django-simple-captcha==0.5.13
django-simple-menu==1.2.1
django-timezone-field==4.2.3
django-tz-detect==0.3.0
django-widget-tweaks==1.4.1
djangorestframework==3.12.4
djangorestframework-queryfields==1.0.0
djangorestframework-xml==2.0.0
djangosaml2==1.3.3
dnspython==2.2.1
drf-yasg==1.21.4
elementpath==3.0.2
email-validator==1.3.0
exceptiongroup==1.0.1
executing==1.2.0
fastapi==0.86.0
flake8==5.0.4
fonttools==4.38.0
future==0.18.2
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.10.0
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.6.0
iso8601==0.1.13
itypes==1.2.0
jarvis-tools==2022.5.20
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
lark==1.1.4
lightgbm==2.2.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mock==2.0.0
mongoengine==0.23.0
mypy-extensions==0.4.3
networkx==2.8.8
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
optimade==0.19.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pathspec==0.10.1
pbr==5.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.3
pluggy==1.0.0
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pydocstyle==6.1.1
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.6.0
pymongo==3.11.2
pyOpenSSL==22.1.0
pyparsing==2.4.7
pysaml2==7.2.1
pytest==7.2.0
pytest-django==4.5.2
python-crontab==2.6.0
python-dateutil==2.8.1
python-dotenv==0.21.0
python3-openid==3.2.0
pytorch-ignite==0.4.10
pytz==2022.6
redis==3.5.3
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==0.22.2
scipy==1.9.3
Sickle==0.7.0
signals-utils==1.14.0
six==1.15.0
sniffio==1.3.0
snowballstemmer==2.2.0
spglib==2.0.2
sqlparse==0.4.3
stack-data==0.6.0
starlette==0.20.4
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.0
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xml-utils==1.18.2
xmlschema==2.1.1
xmltodict==0.12.0
zipp==3.10.0
